webster/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
Pillow==7.1.2
psycopg2==2.8.6
pytz==2020.4
requests==2.23.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
